+++ src/quemb/molbe/be_parallel.py
@@ -8,8 +8,8 @@ from multiprocessing import Pool
 import numpy
 from pyscf import ao2mo, fci, mcscf
 
-from quemb.molbe.external.ccsd_rdm import make_rdm1_uccsd, make_rdm2_uccsd
-from quemb.molbe.external.unrestricted_utils import make_uhf_obj
+from quemb.shared.external.ccsd_rdm import make_rdm1_uccsd, make_rdm2_uccsd
+from quemb.shared.external.unrestricted_utils import make_uhf_obj
 from quemb.molbe.helper import (
     get_eri,
     get_frag_energy,
