+++ src/quemb/molbe/mbe.py
@@ -304,7 +304,7 @@ class BE(MixinLocalize):
     # The following imports turn the imported functions into proper methods
     #  cannot be moved to head of file.
     from quemb.molbe._opt import optimize  # noqa: PLC0415
-    from quemb.molbe.external.optqn import get_be_error_jacobian  # noqa: PLC0415
+    from quemb.shared.external.optqn import get_be_error_jacobian  # noqa: PLC0415
     from quemb.molbe.rdm import compute_energy_full, rdm1_fullbasis  # noqa: PLC0415
 
     def print_ini(self):
