+++ src/quemb/molbe/solver.py
@@ -8,14 +8,14 @@ import numpy
 from pyscf import ao2mo, cc, fci, mcscf, mp
 from pyscf.cc.ccsd_rdm import make_rdm2
 
-from quemb.molbe.external.ccsd_rdm import (
+from quemb.shared.external.ccsd_rdm import (
     make_rdm1_ccsd_t1,
     make_rdm1_uccsd,
     make_rdm2_uccsd,
     make_rdm2_urlx,
 )
-from quemb.molbe.external.uccsd_eri import make_eris_incore
-from quemb.molbe.external.unrestricted_utils import make_uhf_obj
+from quemb.shared.external.uccsd_eri import make_eris_incore
+from quemb.shared.external.unrestricted_utils import make_uhf_obj
 from quemb.molbe.helper import get_frag_energy, get_frag_energy_u, unused
 from quemb.shared import be_var
