+++ src/quemb/shared/external/cpmp2_utils.py
@@ -7,8 +7,8 @@ import numpy as np
 import scipy.linalg as slg
 from pyscf import ao2mo, scf
 
-from quemb.molbe.external.cphf_utils import cphf_kernel_batch as cphf_kernel
-from quemb.molbe.external.cphf_utils import get_cpuhf_u_batch as cpuhf_kernel
+from quemb.shared.external.cphf_utils import cphf_kernel_batch as cphf_kernel
+from quemb.shared.external.cphf_utils import get_cpuhf_u_batch as cpuhf_kernel
 
 """ RMP2 implementation
 """
