+++ src/quemb/shared/external/optqn.py
@@ -8,9 +8,9 @@
 
 import numpy
 
-from quemb.molbe.external.cphf_utils import cphf_kernel_batch, get_rhf_dP_from_u
-from quemb.molbe.external.cpmp2_utils import get_dPmp2_batch_r
-from quemb.molbe.external.jac_utils import get_dPccsdurlx_batch_u
+from quemb.shared.external.cphf_utils import cphf_kernel_batch, get_rhf_dP_from_u
+from quemb.shared.external.cpmp2_utils import get_dPmp2_batch_r
+from quemb.shared.external.jac_utils import get_dPccsdurlx_batch_u
 from quemb.molbe.helper import get_eri, get_scfObj
 from quemb.shared import be_var
