+++ src/quemb/shared/external/restore_eri_addition.c
