+++ src/quemb/shared/external/uccsd_eri.py
